Feedback: helper clas can handle cache.
